% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_script.R
\name{deploy_script}
\alias{deploy_script}
\title{Deploy the MIDI mapper script to a given target directory and make executable}
\usage{
deploy_script(dir = "~", overwrite = FALSE)
}
\arguments{
\item{dir}{The target directory, default: "~" (home directory)}

\item{overwrite}{Whether to overwrite an already existing script at the
target directory, default: FALSE}
}
\value{
The path to the deployed script. You will want to edit and adjust
this script before running it.
}
\description{
Deploy the MIDI mapper script to a given target directory and make executable
}
\examples{
\dontrun{
 target_file <- deploy_script()
}
}
