% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mididrumR-package.R
\docType{package}
\name{mididrumR-package}
\alias{mididrumR}
\alias{mididrumR-package}
\title{mididrumR: Build R Packages By Playing Drums}
\description{
'mididrumR' provides a shell script to listen to MIDI sources, 
    such as electronic drumkits, keyboards, or any other sources, 
    and trigger arbitrary keyboard shortcuts, which can be
    used to load, document, test, build, and check R packages from known IDEs
    such as RStudio, or used in any other context.
}
\author{
\strong{Maintainer}: Florian Mayer \email{Florian.Mayer@dbca.wa.gov.au} (\href{https://orcid.org/0000-0003-4269-4242}{ORCID})

}
\keyword{internal}
